+++ Cargo.toml
@@ -11,7 +11,7 @@ log = { version = "0.4", features = [
     "max_level_debug",
     "release_max_level_warn",
 ] }
-rand = "0.8.5"
+rand = "0.8"
 
 [features]
 default = [
